python/requirements.txt
openai-whisper
anthropic
ffmpeg-python
